Plot Method for unidata Objects

This method allows visualizing `unidata` and `multidata` objects using the `plot.zoo` function from the `zoo` package. It converts the `unidata` (`multidata`) object into a `zoo` object and then applies the plotting function to the time series contained in the object.

Arguments

x

An object of class `unidata` or `multidata`. This object must contain the time series in the `series` slot and the associated time points in the `times` slot.

...

Additional arguments passed to the `plot.zoo` function for customizing the plot (e.g., titles, colors, etc.).